Unify config parsers
